+++ internal/server/api/shared/middleware.go
@@ -48,3 +48,14 @@ func (m *Middleware) IsLicenseExpired() gin.HandlerFunc {
 		}
 	}
 }
+
+func (m *Middleware) OnlyAuthorized() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		_, ok := c.Get(gb.KeyUser)
+		if !ok {
+			c.AbortWithStatusJSON(http.StatusUnauthorized, map[string]string{
+				"message": "Unauthorized user",
+			})
+		}
+	}
+}
