+++ internal/server/api/shared/middleware_test.go
